Don't wait on popup render to push tab onboarding
Rather than do backflips making this work in a React rendering context,
check the Redux store before a render to see if the popup should instead
open a new onboarding tab.

Refs #2025
