+++ ui/index.ts
@@ -1,15 +1,15 @@
-import { newProxyStore } from "@tallyho/tally-background"
 import React, { ComponentType } from "react"
 import ReactDOM from "react-dom"
 import { Store } from "webext-redux"
+import { browser, newProxyStore } from "@tallyho/tally-background"
+import { SUPPORT_TABBED_ONBOARDING } from "@tallyho/tally-background/features"
 import "./_locales/i18n"
 import Popup from "./pages/Popup"
 import Tab from "./pages/Tab"
 
-export { Popup, Tab }
-
-export async function attachUiToRootElement(
-  component: ComponentType<{ store: Store }>
+async function attachUIToRootElement(
+  component: ComponentType<{ store: Store }>,
+  store?: Store
 ): Promise<void> {
   const rootElement = document.getElementById("tally-root")
 
@@ -19,10 +19,40 @@ export async function attachUiToRootElement(
     )
   }
 
-  const backgroundStore = await newProxyStore()
+  const backgroundStore = store ?? (await newProxyStore())
 
   ReactDOM.render(
     React.createElement(component, { store: backgroundStore }),
     rootElement
   )
+}
+
+export async function attachTabUIToRootElement(): Promise<void> {
+  await attachUIToRootElement(Tab)
+}
+
+export async function attachPopupUIToRootElement(): Promise<void> {
+  const store = await newProxyStore()
+
+  if (SUPPORT_TABBED_ONBOARDING) {
+    const state = store.getState()
+    if (Object.keys(state.account?.accountsData?.evm).length === 0) {
+      // we're onboarding! look for an onboarding tab, or open a new one,
+      // rather than rendering the popup
+      const baseURL = browser.runtime.getURL("tab.html")
+      const tabs = (await browser.tabs.query({ url: baseURL })).filter(
+        (tab) => tab?.url && tab.url.includes("onboarding")
+      )
+      if (tabs.length > 0 && tabs[0]?.id) {
+        await browser.tabs.update(tabs[0].id, { active: true })
+      } else {
+        await browser.tabs.create({
+          url: browser.runtime.getURL("tab.html#onboarding"),
+        })
+      }
+      window.close()
+    }
+  }
+
+  await attachUIToRootElement(Popup, store)
 }
