+++ ui/pages/Onboarding/OnboardingInfoIntro.tsx
@@ -1,11 +1,7 @@
-import React, { ReactElement, useEffect, useState } from "react"
+import React, { ReactElement, useState } from "react"
 import { Redirect } from "react-router-dom"
 import { getAddressCount } from "@tallyho/tally-background/redux-slices/selectors"
-import {
-  HIDE_TOKEN_FEATURES,
-  SUPPORT_TABBED_ONBOARDING,
-} from "@tallyho/tally-background/features"
-import { browser } from "@tallyho/tally-background"
+import { HIDE_TOKEN_FEATURES } from "@tallyho/tally-background/features"
 import { useBackgroundSelector } from "../../hooks"
 import SharedButton from "../../components/Shared/SharedButton"
 import SharedProgressIndicator from "../../components/Shared/SharedProgressIndicator"
@@ -87,29 +83,6 @@ export default function OnboardingInfoIntro({
 }: {
   embedded: boolean
 }): ReactElement {
-  useEffect(() => {
-    // if tabbed onboarding is supported, exit the popup and prefer the tab
-    if (SUPPORT_TABBED_ONBOARDING) {
-      const baseURL = browser.runtime.getURL("tab.html")
-
-      const openOnboardingTab = async () => {
-        const tabs = (await browser.tabs.query({ url: baseURL })).filter(
-          (tab) => tab?.url && tab.url.includes("onboarding")
-        )
-        if (tabs.length > 0 && tabs[0]?.id) {
-          await browser.tabs.update(tabs[0].id, { active: true })
-          window.close()
-        } else {
-          await browser.tabs.create({
-            url: browser.runtime.getURL("tab.html#onboarding"),
-          })
-        }
-      }
-
-      openOnboardingTab()
-    }
-  }, []) // empty dependency array means run once and only once
-
   const [activeStep, setActiveStep] = useState(1)
   const [redirectToAddWallet, setRedirectToAddWallet] = useState(false)
